requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
docopt==0.6.2
filelock==3.20.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.12.0
google-auth==2.47.0
google-genai==1.57.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langsmith==0.6.1
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
nodejs-wheel-binaries==24.12.0
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pipreqs==0.4.13
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pytubefix==10.3.6
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.12.0
websockets==15.0.1
xxhash==3.6.0
yarg==0.1.10
yarl==1.22.0
zstandard==0.25.0
